fix: address PR #85 review feedback
Normalize packages/** to packages/* in both the WorkspaceSnapshotReader (git-ref path) and the working-tree snapshot helper so double-star workspace globs are no longer silently reduced to a literal path that skips every child package.
Extract the duplicated gitMergeBase and snapshotFromWorktree helpers from deps-detect and deps-regen into a shared utils/worktree-snapshot.ts module, replacing the non-portable execFileSync ls call with readdirSync along the way.
Guard randomFilename against silent overwrites by checking existsSync before returning a candidate, falling back to a timestamp suffix after 20 unlucky picks — the 1,000-combination triplet space exhausts faster than expected on busy repos.
Update two init-command tests to expect the legacy-versionFiles deprecation note the 0.9.0 deprecation cycle now emits.
Squash the pre-existing pure-dependency changeset into the typed-release-surfaces feature changeset and refresh the dependency table against the cumulative diff from v0.8.0.
