--- .changeset/happy-eagle-2b3316f1.md
@@ -1,11 +1,0 @@
----
-"@savvy-web/changesets": patch
----
-
-## Dependencies
-
-| Dependency | Type | Action | From | To |
-| :--- | :--- | :--- | :--- | :--- |
-| @changesets/cli | peerDependency | updated | ^2.30.0 | ^2.31.0 |
-| @changesets/cli | devDependency | updated | ^2.30.0 | ^2.31.0 |
-| workspaces-effect | dependency | updated | ^0.5.1 | ^0.6.0 |
